add cogs for handlers
